requirements.txt
python-telegram-bot
python-dotenv
pyyaml
watchdog

google-generativeai
IPython
requests